requirements.txt
matplotlib==3.9.1
numpy==2.0.1
Pillow==10.4.0
PyQt6==6.7.1
PyQt6_sip==13.8.0
python-dotenv==1.0.1
Requests==2.32.3
tensorflow==2.17.0
tensorflow_hub==0.16.1
tqdm==4.66.4
